Both graphs seem to be finished

+++ ServerCrow/headers/database.hpp
@@ -105,7 +105,7 @@ public:
 	OrderedProduct getOrderedProductByName(const std::string& name);
 	std::unordered_map<int, OrderedProduct> getOrderedProductsByPage(const int& page);
 	std::unordered_map<int, OrderedProduct> getOrderedProductsByMenu(const std::string& menu);
-	std::unordered_map<int, OrderedProduct> getOrderedProductsByPrice(const int& price);
+	std::unordered_map<int, OrderedProduct> getOrderedProductsByPrice(const double& price);
 	std::unordered_map<int, OrderedProduct> getOrderedProductsByDate(const std::string& date, const std::string& mode);
 
 	std::vector<Ingredient> getIngredients() override;
